Move extensions on TopicDescription to TopicDescriptionExtensions. Do not pollute TopicDescription with responsibilities that don't belong to it.
